single number: approach: Bit Manipulation
If we take XOR of zero and some bit, it will return that bit.
  e.g. a XOR 0 = a
If we take XOR of two same bits, it will return 0.
  e.g. a XOR a = 0
a XOR b XOR a = (a XOR a) XOR b = 0 XOR b = b

So we can XOR all bits togerther to find the unique number.